src/llm_reviewer/requirements.txt
openai
pytest
pytest-cov
jsonschema
nbformat
pandas
fuzzywuzzy
python-dotenv